-Add Demangler.CrackVftable -Add DOSFile, OMFFile, OMBLIBFile -Fix several file types not disposing their globalBlock -Store name on LongImportLibraryMember -Register symbol memory on LongImportLibraryMember, OBJFile -Add NB05, NB09 Symbol Accessors -Rework SymbolMemoryTracker to use a reader/writer lock rather than a monitor -Rework CvDebugSSubsectionHeader -Enhance FileInfo -Fix Modi types not aligning properly -Add Modi50 -Add NMT.Hash -Add config files for automatically generating DIA style extension methods -Read buckets in GSI, add support for querying symbols by name -Rework PDB Map to match native generic implementation properly by using HashClass type -Add publics address map, thunk entry support -Add PDB SrcHeaders support -Add support for reading PDB type numeric data -Add support for reading sourcelink data -Fix MessageResourceEntry not being aligned -Fix several issues with reading VsVersionInfo data -Fix several types not writing their views properly -Add support for reading DNRB symbols -Fix several Merger issues -Improve view writing -Improve Detector logic -Reorganize Locator -Add PooledList

+++ src/PESpy/PE/03. Exception/UnwindCode.cs
@@ -198,7 +198,7 @@ namespace PESpy
             public ushort StackOffset { get; }
 
             public override int StructSize =>
-                sizeof(int) + //StackOffset
+                sizeof(ushort) + //StackOffset
                 base.StructSize;
 
             public SaveXmm128(int offset, byte codeOffset, UnwindInfo.X64Register register, ushort stackOffset) : base(offset, codeOffset, UWOP.SAVE_XMM128)
